+++ RELEASES.md
@@ -9,6 +9,9 @@ Version 1.2.0 (unreleased)
 ==========================
 
   * The crate now requires rustc version 1.31.0 or later.
+  * The implementation details of `gmp::randstate_t` have been changed
+    to reflect that GMP can leave some fields unused and
+    uninitialized.
 
 Version 1.1.14 (2019-07-04)
 ===========================